mixin: restructure package into logical structure
